商品管理

+++ shop/src/com/interview/hobart/shop/entity/ProductInfo.java
@@ -28,7 +28,7 @@ public class ProductInfo implements java.io.Serializable {
 	private Integer id;
 	private CategoryInfo category;
 	private String name;
-	private Integer price;
+	private Double price;
 	private String pic;
 	private String remark;
 	private String xremark;
@@ -55,7 +55,7 @@ public class ProductInfo implements java.io.Serializable {
 
 
 	/** full constructor */
-	public ProductInfo(CategoryInfo category, String name, Integer price, String pic,
+	public ProductInfo(CategoryInfo category, String name, Double price, String pic,
 			String remark, String xremark, Date date, Boolean commend,
 			Boolean open) {
 		this.category = category;
@@ -69,7 +69,7 @@ public class ProductInfo implements java.io.Serializable {
 		this.open = open;
 	}
 
-	public ProductInfo(Integer id, String name, Integer price, String pic,
+	public ProductInfo(Integer id, String name, Double price, String pic,
 			String remark, String xremark, Boolean commend, Boolean open) {
 		super();
 		this.id = id;
@@ -115,11 +115,11 @@ public class ProductInfo implements java.io.Serializable {
 	}
 
 	@Column(name = "price", precision = 8)
-	public Integer getPrice() {
+	public Double getPrice() {
 		return this.price;
 	}
 
-	public void setPrice(Integer price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
